pyproject.toml
[project]
name = "audiointel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.10.2",
    "beautifulsoup4>=4.14.3",
    "black>=26.1.0",
    "chromadb>=1.5.1",
    "email-validator>=2.3.0",
    "fake-useragent>=2.2.0",
    "faker>=40.4.0",
    "fastapi>=0.129.2",
    "flake8>=7.3.0",
    "gotrue>=2.12.4",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "isort>=8.0.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.14",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.0.9",
    "langgraph-checkpoint>=4.0.0",
    "langgraph-prebuilt>=1.0.8",
    "langsmith>=0.7.6",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "passlib[bcrypt]>=1.7.4",
    "playwright>=1.58.0",
    "postgrest>=2.28.0",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "selenium>=4.41.0",
    "starlette>=0.52.1",
    "supabase>=2.28.0",
    "uvicorn[standard]>=0.41.0",
]
